requirements.txt
async_timeout==3.0.1
asyncio==3.4.3
python_dotenv==0.15.0
discord.py==1.6.0
PyNaCl==1.4.0